[BE] refactor: redis 관련 설정 및 cache 관련 컴포넌트 폴더 이동 및 통합 리팩토링 진행
